requirements.txt
ultralytics
opencv-python-headless
pandas
numpy
Pillow
pillow-heif
tqdm
albumentations
matplotlib
jupyter
easyocr
ipykernel
pyyaml
streamlit
plotly
pyngrok
imagehash
pytesseract
streamlit_webrtc
av